# backend/docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: KIWI - API REST v1
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:3333
paths:
...
  '/user/{userId}':
    get:
      tags:
      - User
      summary: Get User
      operationId: getUser
      parameters:
      - $ref: '#/components/parameters/userId'
      responses:
        200:
          $ref: '#/components/responses/userResponse'
        default:
          description: Error
    post:
      tags:
      - User
      summary: Add User to Queue
      operationId: addUserToQueue
      parameters:
       - $ref: '#/components/parameters/userId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
              - ingressCode
              type: object
              properties:
                ingressCode:
                  type: string
                  description: Código de acesso da Queue
            example:
              ingressCode: euS2burrito612
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  position:
                    type: object
                    description: Posição do User na Queue
                    properties:
                      id:
                        type: integer
                      queueId:
                        type: integer
                      userId:
                        type: integer
                      updatedAt:
                        type: string
                      createdAt:
                        type: string
                      first:
                        type: boolean
                        description: Se é a primeira posição na Queue, caso não for, o valor é `null`
                      next:
                        type: integer
                        default: null
                        description: Id da próxima posição, vai ser sempre `null` pois é a última posição
              example:
                position:
                  id: 30
                  queueId: 8
                  userId: 3
                  updatedAt: '2020-11-02T07:38:15.842Z'
                  createdAt: '2020-11-02T07:38:15.842Z'
                  first: null
                  next: null
        default:
          description: Error
  '/user/queue/{queueId}':
    get:
      tags:
      - User
      summary: List Queue's Users
      description: Lista a posição de todos os Users de uma Queue
      operationId: listQueueUsers
      parameters:
      - $ref: '#/components/parameters/queueId'
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    position:
                      type: integer
                    id:
                      type: integer
                    first:
                      type: boolean
                    createdAt:
                      type: string
                    updatedAt:
                      type: string
                    queueId:
                      type: integer
                    next:
                      type: integer
                    userId:
                      type: integer
              example:
                - position: 1
                  id: 20
                  first: true
                  createdAt: '2020-11-02T04:21:32.598Z'
                  updatedAt: '2020-11-02T04:31:40.635Z'
                  queueId: 1
                  next: 21
                  userId: 5
                - position: 2
                  id: 21
                  first: null
                  createdAt: '2020-11-02T04:22:51.627Z'
                  updatedAt: '2020-11-02T04:22:55.265Z'
                  queueId: 1
                  next: null
                  userId: 4
        default:
          description: Error
components:
  parameters:
...
    queueId:
      name: queueId
      in: path
      description: Queue Id
      required: true
      schema:
        type: integer
    userId:
      name: userId
      in: path
      description: User Id
      required: true
      schema:
        type: integer
...
  responses:
...
    userResponse:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              phone:
                type: string
          example:
            id: 4
            name: Nadya
            phone: '994588754'
  securitySchemes:
    httpBearer:
      type: http
      scheme: bearer
security:
- httpBearer: []
...